Quick & dirty 2D physics demo over holiday break
I hadn't messed around seriously with writing physics code for a couple
of years, and wanted to do something simple over the holiday break to
brush up on it. This is based on the sequential impulse solving method
Erin Catto describes in his GDC slides, and after getting stuck on a
pieces (my time over the break was short, and my memory slow to come
back), I ended up borrowing a few pieces from his Box2D_Lite sample to
fill in gaps. This is a very sloppy codebase that was a hobby project
done in a matter of several hours. Bugs galore I'm sure :)